Linked list
A linked list is a data structure consisting of a sequence of elements called "nodes." Each node contains both data and a reference (usually a pointer) to the next node in the sequence. Linked lists are particularly useful for dynamic memory allocation as they can efficiently accommodate varying data sizes during runtime. They do not require contiguous memory allocation like arrays. The first node in the list is called the "head," and the last node's reference points to NULL, indicating the end of the list. Linked lists allow for efficient insertion and deletion of elements, as well as traversal of the data. However, they consume extra memory for the pointers, which is a trade-off for their dynamic nature and constant-time insertions and deletions. Various types of linked lists exist, including singly linked lists, doubly linked lists, and circular linked lists, each with specific features and use cases.